Protegendo rotas

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { UpdateTaskDto } from './dto/update-task.dto';
 
 import { PrismaService } from '../prisma/prisma.service'
 import { PaginationDto } from 'src/common/dto/pagination.dto';
+import { PayloadTokenDto } from 'src/auth/dto/payload-token.dto';
 
 @Injectable()
 export class TasksService {
@@ -64,7 +65,7 @@ export class TasksService {
 
     }
 
-    async createTask(createTaskDto: CreateTaskDto) {
+    async createTask(createTaskDto: CreateTaskDto, tokenPayLoad: PayloadTokenDto) {
 
         try {
 
@@ -73,7 +74,7 @@ export class TasksService {
                     name: createTaskDto.name,
                     description: createTaskDto.description,
                     completed: false,
-                    userId: createTaskDto.userId
+                    userId: tokenPayLoad.sub
                 }
             })
 
@@ -100,7 +101,7 @@ export class TasksService {
         // return newTask
     }
 
-    async update(id: number, updateTaskDto: UpdateTaskDto) {
+    async update(id: number, updateTaskDto: UpdateTaskDto, tokenPayLoad: PayloadTokenDto) {
 
         try {
 
@@ -113,6 +114,11 @@ export class TasksService {
             if (!findTask) {
                 throw new HttpException("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
             }
+
+            if (findTask.userId !== tokenPayLoad.sub) {
+                throw new HttpException("Essa tarefa não pertence ao usuário", HttpStatus.BAD_REQUEST)
+            }
+
 
 
             const task = await this.prisma.task.update({
@@ -156,7 +162,7 @@ export class TasksService {
 
     }
 
-    async delete(id: number) {
+    async delete(id: number, tokenPayLoad: PayloadTokenDto) {
 
         try {
 
@@ -168,6 +174,11 @@ export class TasksService {
 
             if (!findTask) {
                 throw new HttpException("Essa tarefa não existe!", HttpStatus.NOT_FOUND)
+            }
+
+            if (findTask.userId !== tokenPayLoad.sub) {
+                throw new HttpException("Falha ao deletar essa tarefa", HttpStatus.NOT_FOUND)
+
             }
 
             await this.prisma.task.delete({
